Split Metric and MetricType in different files
This makes sure Glean.ts isn't mistakenly imported, as
a circular dependency, when it's not strictly needed.
