+++ glean/src/core/glean.ts
@@ -11,7 +11,8 @@ import { isUndefined, sanitizeApplicationId } from "./utils.js";
 import { CoreMetrics } from "./internal_metrics.js";
 import EventsDatabase from "./metrics/events_database.js";
 import UUIDMetricType from "./metrics/types/uuid.js";
-import DatetimeMetricType, { DatetimeMetric } from "./metrics/types/datetime.js";
+import DatetimeMetricType from "./metrics/types/datetime.js";
+import { DatetimeMetric } from "./metrics/types/datetime_metric.js";
 import Dispatcher from "./dispatcher.js";
 import CorePings from "./internal_pings.js";
 import { registerPluginToEvent, testResetEvents } from "./events/utils.js";
