+++ glean/src/core/metrics/types/boolean_metric.ts
@@ -1,0 +1,19 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+import { Metric } from "../metric.js";
+import { isBoolean } from "../../utils.js";
+
+export class BooleanMetric extends Metric<boolean, boolean> {
+  constructor(v: unknown) {
+    super(v);
+  }
+
+  validate(v: unknown): v is boolean {
+    return isBoolean(v);
+  }
+  payload(): boolean {
+    return this._inner;
+  }
+}
