+++ glean/src/core/metrics/utils.ts
@@ -6,11 +6,11 @@ import { LabeledMetric } from "./index.js";
 import { Metric } from "./metric.js";
 import { JSONValue } from "../utils.js";
 
-import { BooleanMetric } from "./types/boolean.js";
-import { CounterMetric } from "./types/counter.js";
-import { DatetimeMetric } from "./types/datetime.js";
-import { StringMetric } from "./types/string.js";
-import { UUIDMetric } from "./types/uuid.js";
+import { BooleanMetric } from "./types/boolean_metric.js";
+import { CounterMetric } from "./types/counter_metric.js";
+import { DatetimeMetric } from "./types/datetime_metric.js";
+import { StringMetric } from "./types/string_metric.js";
+import { UUIDMetric } from "./types/uuid_metric.js";
 
 /**
  * A map containing all supported internal metrics and its constructors.
