+++ glean/tests/core/metrics/database.spec.ts
@@ -5,7 +5,8 @@
 import assert from "assert";
 
 import Database, { isValidInternalMetricsRepresentation } from "../../../src/core/metrics/database";
-import StringMetricType, { StringMetric } from "../../../src/core/metrics/types/string";
+import StringMetricType from "../../../src/core/metrics/types/string";
+import { StringMetric } from "../../../src/core/metrics/types/string_metric";
 import { JSONValue } from "../../../src/core/utils";
 import Glean from "../../../src/core/glean";
 import { Lifetime } from "../../../src/core/metrics/lifetime";
