+++ glean/tests/core/metrics/datetime.spec.ts
@@ -6,7 +6,8 @@ import assert from "assert";
 import sinon from "sinon";
 
 import Glean from "../../../src/core/glean";
-import DatetimeMetricType, { DatetimeMetric } from "../../../src/core/metrics/types/datetime";
+import DatetimeMetricType from "../../../src/core/metrics/types/datetime";
+import { DatetimeMetric } from "../../../src/core/metrics/types/datetime_metric";
 import TimeUnit from "../../../src/core/metrics/time_unit";
 import { Lifetime } from "../../../src/core/metrics/lifetime";
